requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0

# GIS and training (for encroachment dataset tooling and optional local training)
shapely>=2.0.0
pyproj>=3.5.0
ultralytics>=8.1.0,<9
streamlit>=1.36.0
geopy>=2.4.1
streamlit-drawable-canvas>=0.9.3
pandas>=2.2.0
PyYAML>=6.0

# Auth/JWT (required by backend)
python-jose[cryptography]==3.3.0
cryptography>=42.0.0

# Dataset scraping (for image collection by class)
bing-image-downloader==1.1.2
duckduckgo-search>=6.1.9
requests>=2.31.0

# Kaggle CLI/API for dataset uploads
kaggle>=1.6.14
